refactor: update "cqp-api" dependency in "cqp-storage-elasticsearch" to "0.0.15"
Update API
