+++ cqp-storage-elasticsearch/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 
 group = "com.quantori"
 description = "Chem query platform. Storage Elasticsearch"
-version = "0.0.12"
+version = "0.0.13"
 
 tasks.named<Javadoc>("javadoc") {
     exclude(
@@ -16,7 +16,7 @@ tasks.named<Javadoc>("javadoc") {
 }
 
 dependencies {
-    implementation("com.quantori:cqp-api:0.0.14")
+    implementation("com.quantori:cqp-api:0.0.15")
     implementation("com.quantori:cqp-core:0.0.13")
     implementation("co.elastic.clients:elasticsearch-java:8.6.2")
     implementation(libs.jackson)
