+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchMoleculeUploader.java
@@ -2,8 +2,8 @@ package com.quantori.cqp.storage.elasticsearch;
 
 import co.elastic.clients.elasticsearch.ElasticsearchAsyncClient;
 import co.elastic.clients.elasticsearch.core.bulk.BulkOperation;
+import com.quantori.cqp.api.MoleculesFingerprintCalculator;
 import com.quantori.cqp.api.model.upload.Molecule;
-import com.quantori.cqp.api.service.MoleculesFingerprintCalculator;
 import com.quantori.cqp.core.model.ItemWriter;
 import lombok.extern.slf4j.Slf4j;
