+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchStorageLibrary.java
@@ -14,10 +14,10 @@ import co.elastic.clients.elasticsearch.core.UpdateByQueryRequest;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import co.elastic.clients.transport.ElasticsearchTransport;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.quantori.cqp.api.StorageLibrary;
 import com.quantori.cqp.api.model.Library;
 import com.quantori.cqp.api.model.LibraryType;
 import com.quantori.cqp.api.model.Property;
-import com.quantori.cqp.api.service.StorageLibrary;
 import com.quantori.cqp.storage.elasticsearch.model.LibraryDocument;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.tuple.Pair;
