+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchStorageMolecules.java
@@ -15,15 +15,15 @@ import co.elastic.clients.elasticsearch.core.SearchRequest;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import co.elastic.clients.json.JsonData;
 import co.elastic.clients.transport.ElasticsearchTransport;
+import com.quantori.cqp.api.MoleculesFingerprintCalculator;
+import com.quantori.cqp.api.MoleculesMatcher;
+import com.quantori.cqp.api.StorageMolecules;
 import com.quantori.cqp.api.model.ConjunctionCriteria;
 import com.quantori.cqp.api.model.FieldCriteria;
 import com.quantori.cqp.api.model.Flattened;
 import com.quantori.cqp.api.model.Library;
 import com.quantori.cqp.api.model.Property;
 import com.quantori.cqp.api.model.upload.Molecule;
-import com.quantori.cqp.api.service.MoleculesFingerprintCalculator;
-import com.quantori.cqp.api.service.MoleculesMatcher;
-import com.quantori.cqp.api.service.StorageMolecules;
 import com.quantori.cqp.api.util.FingerPrintUtilities;
 import com.quantori.cqp.core.model.Criteria;
 import com.quantori.cqp.core.model.ExactParams;
